Move TestParquetDataSource to trino-parquet
TestParquetDataSource tests only a generic AbstractParquetDataSource
logic so it should be in trino-parquet.
